# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "Slim Mom OpenAPI 1.0.0",
    "description": "Swagger SlimMom openAPI"
  },
  "servers": [
    {
      "url": "https://slimmom-s41d.onrender.com"
    },
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
    "/api/products/": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Products"
        ],
        "description": "Приватний енд-поінт на отримання денної норми ккал та списку нерекомендованих   продуктів, записує надану/отриману інформацію у БД",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing header with authorization token"
          }
        }
      },
      "get": {
        "tags": [
          "Products"
        ],
        "description": "енд-поінт на пошук продуктів з БД по query-рядку",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing header with authorization token"
          }
        }
      }
    },
    "/api/auth/registration": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/diary/": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/diary/{date}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/diary/{noteId}": {
      "delete": {
        "tags": [
          "Diary"
        ],
        "description": "Eнд-поінт для видалення з`їденого продукту в конкретний день",
        "parameters": [
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing header with authorization token"
          }
        }
      }
    }
  },
  "components": {
    "requestBodies": {
      "userParams": {
        "type": "object",
        "properties": {
          "height": {
            "type": "number",
            "format": "float",
            "minimum": 10,
            "maximum": 300,
            "example": "150"
          },
          "age": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200,
            "example": "18"
          },
          "currentWeight": {
            "type": "integer",
            "minimum": 10,
            "maximum": 500,
            "example": "60"
          },
          "desiredWeight": {
            "type": "integer",
            "minimum": 10,
            "maximum": 500,
            "example": "50"
          },
          "bloodType": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4,
            "example": "1"
          }
        },
        "required": [
          "height",
          "age",
          "currentWeight",
          "desiredWeight",
          "bloodType"
        ],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}